Refactor seek bar preview functionality into a separate file
